+++ games/NEWS.md
@@ -2,6 +2,8 @@
 
 * Fixed bugs in `egame12()` and `egame122()` that caused them to fail when the left-hand side was constructed with binary variables of class `"integer"` (thanks to Tyson Chatagnier for the bug report)
 
+* Streamlined `Mode()` function according to the Stack Overflow answer <http://stackoverflow.com/a/8189441/143383>
+
 
 # games 1.1-1